fixed description_array nil

+++ app/models/exercise.rb
@@ -47,19 +47,21 @@ class Exercise < ActiveRecord::Base
   end
 
   def add_descriptions(description_array)
-    description_array.each do |key, array|
-      destroy = array[:_destroy]
-      id = array[:id]
-      
-      if id
-        description = Description.find(id)
-        if destroy
-          description.destroy
+    if description_array
+      description_array.each do |key, array|
+        destroy = array[:_destroy]
+        id = array[:id]
+        
+        if id
+          description = Description.find(id)
+          if destroy
+            description.destroy
+          else
+            description.update(text: array[:text])
+          end
         else
-          description.update(text: array[:text])
+          descriptions << Description.create(text: array[:text]) unless destroy
         end
-      else
-        descriptions << Description.create(text: array[:text]) unless destroy
       end
     end
   end
